refactor(whispering): unify recording overlay ownership
Move the reveal signal and listener under the recording-overlay event and runtime owners. This removes two one-purpose files while keeping the panel least-privilege and preserving body-click reveal.
