+++ apps/whispering/src/lib/recording-overlay/events.ts
@@ -100,6 +100,9 @@ export const recordingOverlayAction = defineWindowEvent<RecordingOverlayAction>(
 	'recording-overlay:action',
 );
 
+/** overlay -> main: reveal the main Whispering window. */
+export const revealMainWindow = defineWindowSignal('main-window:reveal');
+
 /**
  * overlay -> main: the overlay mounted and its listener is live, so the main
  * window should re-send the latest status. Without this handshake the first
