+++ apps/whispering/src/routes/(app)/_runtime/attach-recording-overlay.svelte.ts
@@ -1,13 +1,22 @@
 import type { UnlistenFn } from '@tauri-apps/api/event';
+import { getCurrentWindow } from '@tauri-apps/api/window';
 import { recordingOverlay } from '#platform/recording-overlay';
 import { tauri } from '#platform/tauri';
-import { recordingOverlayAction } from '$lib/recording-overlay/events';
+import {
+	recordingOverlayAction,
+	revealMainWindow,
+} from '$lib/recording-overlay/events';
 import { dispatchPillAction } from '$lib/recording-overlay/pill-actions';
 import { projectLifecycleToStatus } from '$lib/recording-overlay/projection';
 import { dictationLifecycle } from '$lib/state/dictation-lifecycle.svelte';
 
 export function attachRecordingOverlay() {
-	let unlistenAction: UnlistenFn | undefined;
+	const unlisteners: UnlistenFn[] = [];
+	let destroyed = false;
+	const trackUnlistener = (unlisten: UnlistenFn) => {
+		if (destroyed) unlisten();
+		else unlisteners.push(unlisten);
+	};
 
 	const overlayStatus = $derived(
 		projectLifecycleToStatus(dictationLifecycle.current),
@@ -18,14 +27,24 @@ export function attachRecordingOverlay() {
 	});
 
 	if (tauri) {
-		void (async () => {
-			unlistenAction = await recordingOverlayAction.listen((event) =>
+		void recordingOverlayAction
+			.listen((event) =>
 				dispatchPillAction(event.payload),
-			);
-		})();
+			)
+			.then(trackUnlistener);
+		void revealMainWindow
+			.listen(async () => {
+				const mainWindow = getCurrentWindow();
+				await mainWindow.show();
+				await mainWindow.unminimize();
+				// setFocus often fails on macOS; ignore.
+				await mainWindow.setFocus().catch(() => {});
+			})
+			.then(trackUnlistener);
 	}
 
 	return () => {
-		unlistenAction?.();
+		destroyed = true;
+		for (const unlisten of unlisteners) unlisten();
 	};
 }
